Split express setup into multiple routes. Fix eslint setup
