+++ package.json
@@ -33,6 +33,7 @@
     "dotenv": "^6.2.0",
     "express": "^4.16.4",
     "express-session": "^1.15.6",
+    "express-validation": "^1.0.2",
     "helmet": "^3.15.0",
     "isomorphic-loader": "^2.0.2",
     "joi": "^14.3.1",
@@ -60,6 +61,8 @@
     "babel-loader": "^8.0.4",
     "babel-plugin-async-to-promises": "^1.0.5",
     "eslint": "^5.10.0",
+    "eslint-config-airbnb-base": "^13.1.0",
+    "eslint-plugin-import": "^2.15.0",
     "file-loader": "^3.0.1",
     "mocha": "^4.1.0",
     "nyc": "^13.1.0",
