+++ src/modules/validation.js
@@ -1,20 +1,24 @@
 const Joi = require('joi');
 
-const authResponseSchema = Joi.object().keys({
-  access_token: Joi.string().required(),
-  scope: Joi.string().required(),
-  team_name: Joi.string().required(),
-  team_id: Joi.string().required(),
-  bot: Joi.object().keys({
-    bot_user_id: Joi.string().required(),
-    bot_access_token: Joi.string().required()
-  }).required(),
-  createdAt: Joi.date().default(Date.now())
-}).required();
-
-
 module.exports = {
-  auth: {
-    validate: data => Joi.validate(authResponseSchema, data)
+  slackBotCallbackRaw: {
+    query: {
+      code: Joi.string().required(),
+      state: Joi.string().alphanum().required(),
+      error: Joi.string().allow('')
+    }
+  },
+  slackBotCallbackModified: {
+    auth: {
+      access_token: Joi.string().required(),
+      scope: Joi.string().required(),
+      team_name: Joi.string().required(),
+      team_id: Joi.string().required(),
+      bot: Joi.object().keys({
+        bot_user_id: Joi.string().required(),
+        bot_access_token: Joi.string().required()
+      }).required(),
+      createdAt: Joi.date().default(Date.now())
+    }
   }
 };
