+++ src/routes/dashboard.route.js
@@ -1,0 +1,7 @@
+const express = require('express');
+
+const router = express.Router();
+
+router.get('/', (req, res) => { res.status(200).json('OK'); });
+
+module.exports = router;
